+++ client/src/app/admin/admin-maps/admin-maps.component.ts
@@ -2,6 +2,7 @@ import { Component, inject } from '@angular/core';
 import { AdminService } from '../../_services/admin.service';
 import { RouterLink } from '@angular/router';
 import { TooltipModule } from 'ngx-bootstrap/tooltip';
+import { ConfirmService } from '../../_services/confirm.service';
 
 @Component({
   selector: 'app-admin-maps',
@@ -12,6 +13,7 @@ import { TooltipModule } from 'ngx-bootstrap/tooltip';
 })
 export class AdminMapsComponent {
   adminService = inject(AdminService);
+  private confirmService = inject(ConfirmService);
 
   ngOnInit(): void {
     if (this.adminService.maps().length == 0) this.loadMaps();
@@ -22,6 +24,12 @@ export class AdminMapsComponent {
   }
 
   deleteMap(mapId: number){
-    this.adminService.deleteMap(mapId);
+    this.confirmService.confirm()?.subscribe({
+      next: result => {
+        if (result){
+          this.adminService.deleteMap(mapId);
+        }
+      }
+    })
   }
 }
